pyproject.toml
[project]
name = "gmail-mcp"
version = "0.1.0"
description = "Multi-account Gmail MCP server for Claude"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
gmail-mcp = "server:main"
gmail-mcp-cli = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
